Update to the basic functionality

+++ db/migrate/20150219153513_create_transactions.rb
@@ -2,11 +2,11 @@ class CreateTransactions < ActiveRecord::Migration
   def change
     create_table :transactions do |t|
       t.integer :cart_id, null: false, unique: true
-      t.decimal :subtotal
-      t.decimal :tax
-      t.decimal :total
-      t.decimal :amount_given
-      t.decimal :change
+      t.float :subtotal
+      t.float :tax
+      t.float :total
+      t.float :amount_given
+      t.float :change
       t.timestamps null: false
     end
   end
